+++ src/main/java/dev/embeddings4j/VectorIndex.java
@@ -3,6 +3,6 @@ package dev.embeddings4j;
 import java.util.List;
 
 public interface VectorIndex {
-    void addItem(int index, float[] values);
-    List<Integer> searchNearest(float[] values, int numResults);
+    void addItem(int index, double[] values);
+    List<Integer> searchNearest(double[] values, int numResults);
 }
