+++ src/main/java/dev/embeddings4j/internal/BruteForceSearchVectorIndex.java
@@ -8,28 +8,28 @@ import java.util.PriorityQueue;
 
 public class BruteForceSearchVectorIndex implements VectorIndex {
     private final int vectorDimension;
-    private final List<float[]> vectors;
+    private final List<double[]> vectors;
 
     public BruteForceSearchVectorIndex(int vectorDimension) {
         this.vectorDimension = vectorDimension;
         this.vectors = new ArrayList<>();
     }
 
-    public void addItem(int index, float[] values) {
+    public void addItem(int index, double[] values) {
         if (values.length != vectorDimension) {
             throw new IllegalArgumentException("Vector dimensions do not match");
         }
         vectors.add(index, values);
     }
 
-    public List<Integer> searchNearest(float[] values, int numResults) {
+    public List<Integer> searchNearest(double[] values, int numResults) {
         if (values.length != vectorDimension) {
             throw new IllegalArgumentException("Vector dimensions do not match");
         }
 
-        PriorityQueue<DistanceIndexPair> maxHeap = new PriorityQueue<>((a, b) -> Float.compare(b.distance, a.distance));
+        PriorityQueue<DistanceIndexPair> maxHeap = new PriorityQueue<>((a, b) -> Double.compare(b.distance, a.distance));
         for (int i = 0; i < vectors.size(); i++) {
-            float distance = euclideanDistance(values, vectors.get(i));
+            double distance = euclideanDistance(values, vectors.get(i));
             maxHeap.offer(new DistanceIndexPair(distance, i));
 
             if (maxHeap.size() > numResults) {
@@ -45,20 +45,20 @@ public class BruteForceSearchVectorIndex implements VectorIndex {
         return nearestIndices;
     }
 
-    private float euclideanDistance(float[] vector1, float[] vector2) {
-        float sum = 0;
+    private double euclideanDistance(double[] vector1, double[] vector2) {
+        double sum = 0;
         for (int i = 0; i < vector1.length; i++) {
-            float diff = vector1[i] - vector2[i];
+            double diff = vector1[i] - vector2[i];
             sum += diff * diff;
         }
-        return (float) Math.sqrt(sum);
+        return Math.sqrt(sum);
     }
 
     private static class DistanceIndexPair {
-        float distance;
+        double distance;
         int index;
 
-        DistanceIndexPair(float distance, int index) {
+        DistanceIndexPair(double distance, int index) {
             this.distance = distance;
             this.index = index;
         }
